quick patch to make it work with 3.7 based dropbox
This is a quick patch to make it all work again with the Python 3.7
based builds of Dropbox. There are still more errors thrown when
decompiling than I would like. But at least the opcode database seems to
be constructed properly now and there definitely is some code properly
decompiled. Some more investigation is needed and possibly some work on
the uncompyle6 internals to get better results. It could also be that
I'm simply calling/using uncompyle6 wrong but right now I don't have the
cycles to investigate. At least this commit makes the bitrot of this
repo a bit less worse as it's now functional again.