improve opcode mapping generation
The opcode mapping had a bug in that the newer python versions always
have 2 bytes opcode with the seocnd byte being the argument; we now skip
over those properly to derive the opcode database. In the NOTES file
can be seen that f.e. BUILD_CONST_KEY_MAP is a relatively recent opcode
which was not derived properly and as such uncompyle6 failed to
decompile a lot of the code.

With the new included opcode database or the new opcode database
generation script the results go from a having a successrate of 32%
to a successrate of 92%. Quite an improvement!

before:
__main__ - INFO - Processed 4180 files (1331 succesfully decompiled, 2849 failed)

after:
__main__ - INFO - Processed 4173 files (3825 succesfully decompiled, 348 failed)