easier debugging by dumping the opcode database
make it easier to debug when things break by having this little
checkdb script which dumps out the database and the currently
defined mappings for the specific opcodes.

Thanks to @slinkinone for helping me debug some of the issues
that let to the code-rot of this repository and an erroneous
opcode mapping being generated.